add findUser by ID

+++ src/main/java/com/rouletteapp/roulette/repository/IUserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface IUserRepository extends JpaRepository<User, Integer> {
+public interface IUserRepository extends JpaRepository<User, Long> {
 
 }
